[Son]: switch from double to big decimal

+++ src/main/java/com/example/cabonerfbe/dto/FactorDto.java
@@ -2,6 +2,7 @@ package com.example.cabonerfbe.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @NoArgsConstructor
@@ -13,5 +14,5 @@ public class FactorDto {
     UUID id;
     ImpactMethodCategoryDto impactMethodCategory;
     String scientificValue;
-    double decimalValue;
+    BigDecimal decimalValue;
 }
