+++ src/main/java/com/example/cabonerfbe/dto/UnitDto.java
@@ -2,6 +2,7 @@ package com.example.cabonerfbe.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @NoArgsConstructor
@@ -12,7 +13,7 @@ import java.util.UUID;
 public class UnitDto {
     private UUID id;
     private String name;
-    private double conversionFactor;
+    private BigDecimal conversionFactor;
     private Boolean isDefault;
     private UnitGroupDto unitGroup;
