+++ src/main/java/com/example/cabonerfbe/models/SubscriptionType.java
@@ -4,6 +4,8 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -16,7 +18,7 @@ public class SubscriptionType extends Base{
     private String description;
     private int projectLimit;
     private int usageLimit;
-    private double annualCost;
-    private double monthlyCost;
+    private BigDecimal annualCost;
+    private BigDecimal monthlyCost;
     private boolean canCreateOrganization;
 }
