+++ src/main/java/com/example/cabonerfbe/response/CreateElementaryResponse.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -22,7 +23,7 @@ public class CreateElementaryResponse {
     private String description;
     private LifeCycleStageDto lifeCycleStage;
     private UUID projectId;
-    private double overallProductFlowRequired;
+    private BigDecimal overallProductFlowRequired;
     private List<ProcessImpactValueDto> impacts;
     private List<ExchangesDto> exchanges;
 }
